# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Meal planner backend",
        "title": "Meal Planner API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/meal-plan": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-plan"
                ],
                "summary": "List all meal plan entries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.MealPlanEntryResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-plan"
                ],
                "summary": "Add a meal plan entry",
                "parameters": [
                    {
                        "description": "Entry",
                        "name": "entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.CreateMealPlanEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api.MealPlanEntryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "meal-plan"
                ],
                "summary": "Delete all meal plan entries (reset)",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/meal-plan/day/{dayIndex}": {
            "delete": {
                "tags": [
                    "meal-plan"
                ],
                "summary": "Delete all entries for a specific day and reindex subsequent days",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Day Index (\u003e= 1)",
                        "name": "dayIndex",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/meal-plan/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-plan"
                ],
                "summary": "Update a meal plan entry's serving grams or amount",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Entry",
                        "name": "entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.UpdateMealPlanEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.MealPlanEntryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "meal-plan"
                ],
                "summary": "Remove a single meal plan entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/meal-plan/{id}/customize": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-plan"
                ],
                "summary": "Snapshot recipe ingredients for customization",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.MealPlanEntryIngredientResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.CreateMealPlanEntryRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "day_index": {
                    "type": "integer"
                },
                "ingredient_id": {
                    "type": "integer"
                },
                "person_id": {
                    "type": "integer"
                },
                "recipe_id": {
                    "type": "integer"
                },
                "serving_grams": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.MealPlanEntryResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "customized_ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.MealPlanEntryIngredientResponse"
                    }
                },
                "day_index": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "ingredient": {
                    "$ref": "#/definitions/internal_api.IngredientInEntry"
                },
                "person_id": {
                    "type": "integer"
                },
                "recipe_id": {
                    "type": "integer"
                },
                "serving_grams": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.UpdateMealPlanEntryRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "serving_grams": {
                    "type": "number"
                }
            }
        },
...
    }
}